backend/requirements.txt
# FastAPI and ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==1.8.2

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.7.0
email-validator==2.1.0

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Database - Async SQLAlchemy + PostgreSQL
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# Vector database
qdrant-client==1.12.1

# OpenAI
openai==1.58.1

# PDF processing
pypdf==5.1.0
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13

# Data processing
pandas==2.2.3
openpyxl==3.1.5

# File uploads
python-multipart==0.0.17

# Token counting (for chunking)
tiktoken==0.8.0

# HTTP client
httpx==0.28.1

# Observability - Logging
loguru==0.7.3

# Observability - Metrics
prometheus-client==0.21.0

# Observability - OpenTelemetry (optional)
# Uncomment for OpenTelemetry support:
# opentelemetry-api==1.28.0
# opentelemetry-sdk==1.28.0
# opentelemetry-exporter-otlp==1.28.0
# opentelemetry-instrumentation-fastapi==0.49b0

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Type hints
typing-extensions==4.12.2
